requirements.txt
ollama>=0.1.0
groq>=0.4.0
openai>=2.0.0
requests>=2.31.0
pydub>=0.25.1
pillow>=10.0.0
python-dotenv>=1.0.0
gtts>=2.5.0
torch>=2.0.0
diffusers>=0.35.0
transformers>=4.35.0
accelerate>=1.10.0
edge-tts>=6.1.12
numpy>=1.24.0,<2.0.0
customtkinter>=5.2.0
psutil>=5.9.0
json5>=0.9.0

